requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
ChatterBot==0.8.4
chatterbot-corpus==1.1.1
cycler==0.9.0
decorator==4.0.6
Django==1.10
django-bootstrap3==9.1.0
django-filter==1.1.0
django-mail-templated==2.6.5
django-password-reset==1.0
django-paypal==0.2.5
django-phonenumber-field==1.3.0
django-registration-redux==1.10
django-widget-tweaks==1.4.1
djangorestframework==3.7.3
f90nml==1.0.2
future==0.16.0
get==0.0.39
googlemaps==2.5.1
googletrans==2.2.0
idna==2.6
joblib==0.9.4
lockfile==0.12.2
lxml==4.2.0
mathparse==0.1.1
mechanize==0.2.5
nltk==3.2.5
nose==1.3.7
numpy==1.14.2
oauthlib==2.0.7
olefile==0.44
payu==0.5.1
Pillow==4.3.0
post==0.0.26
psycopg2==2.7.4
public==0.0.65
py==1.5.3
pymongo==3.6.1
pyparsing==2.0.3
PySocks==1.6.8
python-dateutil==2.6.1
python-google-places==1.4.1
python-http-client==3.0.0
python-twitter==3.4.1
PyYAML==3.12
query-string==0.0.28
RapidConnect==0.0.6
request==0.0.26
requests==2.18.4
requests-oauthlib==0.8.0
selenium==3.9.0
sendgrid==5.3.0
simplejson==3.8.1
six==1.11.0
South==1.0.2
SQLAlchemy==1.2.5
stem==1.6.0
torrequest==0.1.0
urllib3==1.22
virtualenv==15.1.0
websocket-client==0.47.0
Babel==2.5.3
dj-database-url==0.5.0
gunicorn==19.7.1
html5lib==1.0.1
httplib2==0.11.3
phonenumberslite==8.9.4
PyPDF2==1.26.0
pytz==2018.4
reportlab==3.4.0
six==1.11.0
webencodings==0.5.1
whitenoise==3.3.1
xhtml2pdf==0.2.2
